backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
Flask-JWT-Extended==4.4.4
python-dotenv==0.19.2
celery==5.2.7
redis==4.3.4
python-dateutil==2.8.2
python-magic==0.4.27
importlib-metadata==4.13.0
Pillow==11.2.1
numpy==2.2.
mpi4py==4.0.3
cv==1.0.0